Sanitize HTTP exporter cgroup inputs (#352)
## Summary
- allow HTTP exporter cgroup setters to accept float inputs and clamp
invalid values
- convert shaper cgroup metric reporting to pass float values to the
exporter
- add ServeHTTP coverage for NaN/negative cgroup metrics and toggling
the unlimited flag

## Testing
- make lint-fix
- make test
- make coverage MIN_COVERAGE=96
- make check


------
[Codex
Task](https://chatgpt.com/codex/tasks/task_e_692c750debc88322afb0ce072acb89fb)

+++ cmd/shaper/cgroup_inspection.go
@@ -68,26 +68,26 @@ func recordCgroupMetrics(exporter *metricshttp.Exporter, info *cgroup.CPU) {
 		return
 	}
 
-	weight := uint64(0)
+	weight := float64(0)
 	if info != nil && info.Weight.Err == nil && info.Weight.Available {
-		weight = info.Weight.Value
+		weight = float64(info.Weight.Value)
 	}
 
 	exporter.SetCgroupCPUWeight(weight)
 
 	var (
-		quota  uint64
-		period uint64
+		quota  float64
+		period float64
 	)
 
 	unlimited := false
 
 	if info != nil && info.Max.Err == nil && info.Max.Available {
-		period = info.Max.Period
+		period = float64(info.Max.Period)
 
 		unlimited = info.Max.Unlimited
 		if !info.Max.Unlimited {
-			quota = info.Max.Quota
+			quota = float64(info.Max.Quota)
 		}
 	}
